broker/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
httpx==0.28.1
python-dotenv==1.0.1
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-api-python-client==2.157.0
google-analytics-data==0.18.16
google-ads==25.2.0
